Convert Value.Bytes to method (#4613)
This is preparation for storing some primitive values natively.  With
that change, Value.bytes won't contain the ZNG encoding for native
values, so this method will have to create it.  That'll be relatively
expensive, but we'll avoid it in most cases by replacing the
zed.DecodeXXX(val.Bytes()) pattern with new Value methods that access a
native value directly when available.