Use native representation for Boolean zed.Value (#4623)
The bytes field of a non-null zed.Value always contains the value's ZNG
representation.  This is wasteful for primitive types with a
fixed-length native representation like Booleans and numeric types
because it consumes memory and requires work to convert between the
native and ZNG representations.  To address that waste, this change
introduces a native representation for Boolean zed.Values.  (Subsequent
changes will extend it to numeric types.)

The basic idea here is to use the unsafe package to store the bits of a
fixed-length native value in the capacity field of Value.bytes's slice
header.  Additionally, a sentinel address (belonging to the package
variable called nativeBase) is stored in the base address field of
Value.byte's slice header to distinguish native Values from ZNG Values.
When the ZNG representation for a native Value is needed, the
Value.Bytes method creates it.

Note that changes to the Go runtime may break this implementation!