fix zson marshaling of unknown interface values (#4633)
This commit changes the logic that determines Zed types from
Go native types to always use zed type null for interface types
since the underlying concrete type is not known.  Most of the
marshaling logic uses values to infer types rather than type
lookups but when there is a nil value of a complex type with
embedded interfaces, this heuristic is needed.  This still
doesn't cover all the possible cases here but it is an improvement
over what was here before.