Parallelize more source paths (#4785)
When looking for source paths to parallelize, the optimizer.walkEntries
examines only the first operator in a sequence.  This prevents the
optimizer from parallelizing a pool scan in the right input of a join
expressed in subquery syntax, like

    from left | join (from right | ...) on left_key=right_key

Change walkEntries to examine all operators in a sequence so almost all
source paths are parallelized.  (The exceptions are paths inside
dag.Over and paths containing no operator after the source.)