lakeapi: Use pool name/reference over pool id
This commit changes the lakeapi to use the pool name (which call also be
the string version of the pool ID) over requiring the pool id when
operating on a given pool. Apart from reducing the amount of requests
needed when operating on a remote lake, this also allows for users to
create a pool with ksuid name with Zed knowing how to resolve this
correctly.

Closes #4431