Replace zed.Value.bytes slice with base and len
A zed.Value occupies five words: two for the Type field (an interface)
and three for the bytes field (a slice).  Since the slice capactiy field
is rarely used (only in Value.CopyFrom), shrink Values to four words by
replacing the bytes field with base (a pointer) and len (a uint64)
fields.  This can reduce maximum RSS by 1.4x when sorting a large input
or loading a large input into a lake.