Fix NaN handling in expr.LookupCompare (#4835)
* Fix NaN handling in expr.LookupCompare

LookupCompare fails to account for the fact that a floating-point NaN is
neither less than, greater than, nor equal to itself or any other
floating-point value, leading to unexpected results from the sort
operator and other parts of the system that compare values.  A common
approach to this quirk is to consider NaN to be equal to itself and
either less than or greater than all other floating-point values.  Apply
this approach with NaN considered less than all other values, in keeping
with the Go standard library.

* Order NaNs