Fix signed zero handling in expr.LookupCompare
LookupCompare fails to account for the fact that a floating-point zero
is signed.  As a result, a ZNG set containing both negative and positive
zero does not have a canonical form.  Fix by ordering negative zero
before positive zero.  The implementation changes the ordering of NaNs
whose sign bit is set but that shouldn't matter.