Limit VNG files to a single segment
The VNG dictionary representation does not work for multiple segments if
the index of a dictionary entry changes after a segment containing that
entry is written.  Work around this by limiting VNG files to a single
segment.

We'll fix this permanently either by removing segments entirely or by
adding per-segment dictionaries.