Replace kernel.Reader with dag.DefaultScan
compiler/kernel.Reader is a little peculiar both because it's the only
compiler/ast/dag.Op implementation outside the dag package and because
it holds zio.Readers.  Separately, it forces a dependency in
compiler/optimizer on compiler/kernel that's getting in my way on vector
work.

Replace it with dag.DefaultScan, which represents the same thing (a scan
of an input stream provided by the runtime) but can live in the dag
package because it doesn't hold zio.Readers (allowing it to be
serialized).  The zio.Readers are instead passed to compiler.Job.Build
and then to kernel.Builder.Build.