simplify and improve VNG format (#4984)
This commit simplifies the VNG format with a new VNG metadata object
called Variant, which represents a sequence of typed values (without
putting the types into a union), in line with the Zed data model.
This replaces the ad hoc data structures that were encoded in the
VNG trailer.  We also changed vectors so they are all represented in
a single segment in the data section, which simplifies vector stats and
reading.  To support streaming, we replaced the trailer with
a header and added framing so that VNG objects can simply be concatenated
in a stream or file.  We added the command "zed dev vng" to displace
the fixed VNG framing header (with magic) and the metadata as Zed data.

Nulls are now handled completely by the Nulls metadata structure and
vector metadata now includes length information, which will be used by the
vcache in a subsequent commit to faciliate easier flattening of leaf
vectors so that all leaf values within a type all have the same length
(except perhaps for vectors below unions, TBD).

The VNG write path now stores each object entirely in memory.  Once
all writes have completed, the vectors are encoded in parallel (e.g.,
applying compression). Then a DFS traversal computes the data layout
and produces a single Zed value representing the vector metadata
with data locations.  The header, metadatasection, and data vectors
are  then written out sequentially from a single thread such that the
buffers end up being written in the same locations that were computed
in the metadata pass.