remove unused logger from op.Context
This commit removes the logger from the query runtime context.
We no longer emit logging info from a query and instead favor
a model where query errors are embedded in the query's output and
system errors are caught with panic and logged.

The task in #4487 will further obviate the need for any logging
from the query runtime.