align vector and sequence runtimes for easier coexistence (#5005)
This commit rearranges the runtime codebase so that the
sequence and vector runtimes coexist in parallel layout.
We added the path runtime/sam for the sequential op and expr
packages etc. and created op and expr packages for runtime/vam.

We moved op.Context to runtime.Context since it is shared
between vam and sam.  To do so and avoid import loops, we moved
runtime.Query and DeleteQuery into runtime/exac where they
more naturally belonged and made interfaces to these implementations.

We also removed runtime/vam/index.go since it won't be needed.