add vector compiler skeleton and basic logic (#5013)
This commit adds basic comparison logic to the vector runtime
and some initial compiler framework to translate vectorized DAGs
more generally into runtime ops and exprs.  This is partially complete
but we want to merge as is so the team can begin work on vector-enabled
field search in parallel.

We also renamed the "zed dev vcache" command to "zed dev vector".

In a subsequent PR, we will finish out the mechanisms for handling
heterogeneously typed data with mixed-in errors building up a library
of helper functions for Zed variants to do so.  At that point, we will
begin adding comprehensive tests for the vector runtime.