Allow pools to be given a ksuid name
The commit fixes an issue with pool looks that made it so pools with a
ksuid-like name could not be looked up for most zed commands. Fix the
logic for lake/api.PoolID so that looking up a pool by ID is tried first
and then falls back to looking up a pool by name if nothing is found.

Closes #4431