Implement vector record expressions (#5273)
Unlike the implementation in runtime/sam/expr/values.go, this one does
not include an optimization for record expressions that do not include
the spread operator.  That's on the assumption it isn't needed here
since it only avoids work once per vector (compared with once per Zed
value over in runtime/sam/expr).