allow panics inside PEG parser (#5301)
This commit changes the default behavior of the PEG parser
to not recover from panics.  Previously, a panic would be caught
and converted to an error then we would report the error as
a syntax error. There should be no panics in the grammar and
if there are, they should be displayed to users so bugs can
be filed and fixed.