change variant to dynamic
This commit changes our terminology for a top-level dynamic
type to "Dynamic" from "Variant".  The term "variant" has been
corrupted by the industry to mean efficiently-encoded
semi-structured data when the traditional meaning from
programming languages is a tagged union.  To avoid this
confusion, we're following clickhouse's terminology and using
the term dynamic instead.  This change is internal to the code
only and does not affect any user-facing documentation as
the super-structured data model is strongly typed throughout
and does not expose the implied top-level dynamic type as
a user-visible type.