Fix panic on filepath.Walk error in ztest.findZTests
If filepath.Walk gets an error from os.Lstat, it invokes its callback
with a nil os.FileInfo and the error.  This will cause findZTests to
panic because its filepath.Walk callback does not check for an error
before calling os.FileInfo.IsDir.  Fix by checking for an error in the
callback.